requirements.txt
openai
azure-functions
azure-storage-blob
azure-search-documents
azure-identity 
azure-core
python-dotenv
pandas
openpyxl
xlrd
pyodbc
langchain-openai